pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cymphony"
version = "0.1.0"
description = "Orchestration service for Claude Code agents driven by Linear issues"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "aiohttp>=3.9",
    "watchdog>=4.0",
]

[project.scripts]
cymphony = "cymphony.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cymphony"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
